feature. change sendFn in store to private method. Move wilddogPaths from constants to store.
