+++ lib/store/index.js
@@ -26,9 +26,9 @@ export default class Store {
     if (conf.clean) {
       connect.syncRef.remove();
     }
-    this.data[key] = new CompareSender(this.sendFn.bind(this, key), initData, conf);
+    this.data[key] = new CompareSender(this._sendFn.bind(this, key), initData, conf);
   }
-  sendFn(key, data) {
+  _sendFn(key, data) {
     const connect = this.connects[key];
     connect.syncRef[connect.setDataMethod](data);
   }
