+++ lib/store/wilddogPaths.js
