+++ lib/store/wilddogconfig.js
@@ -1,4 +1,4 @@
-import getPaths from '../constants/wilddogPaths';
+import getPaths from './wilddogPaths';
 import * as _ from '../utils/lodash';
 
 const config = {
